addition / and reproject of files
