+++ app/services/letter_service.py
@@ -1,14 +1,14 @@
 from services.vector_database import pinecone_delete_namespace
-from services.letter import write_letter_character
+from services.letter.write import write_letter
 from query.letter import Letter
-from query.letter import get_letter_by_id
-from services.embeddings import embed_letter, embed_letter_pinecone
+from query.letter import get_a_letter
+from services.embeddings import embed_letter_pinecone
 from datetime import datetime
 from schemas.schemas import LetterDto
 from models.database import SessionLocal
 
 
-def write_letter(letter: LetterDto):
+def create_letter(letter: LetterDto):
 
     # pinecone_delete_namespace() # 테스트용 코드
     
@@ -21,15 +21,7 @@ def write_letter(letter: LetterDto):
     letter_sending.letter_content =  letter.letter_content
     letter_sending.created_time = datetime.now()
     
-    letter_received = write_letter_character(letter_sending)
-    # letter_receiving = Letter()
-    # letter_receiving.character_id = letter.character_id
-    # letter_receiving.user_id = letter.user_id
-    # letter_receiving.reception_status = "receiving"
-#     letter_receiving.letter_content = """**Dear Inji,**
-# Hermione Granger"""
-    # letter_receiving.created_time = datetime.now()
-
+    letter_received = write_letter(letter_sending)
     
     save_letter(letter_sending)
     save_letter(letter_received)
@@ -42,8 +34,12 @@ def save_letter(letter: Letter):
     try:
         with SessionLocal() as session:
             # embed_letter(letter)
+            
+            # DB에 저장 (insert)
             session.add(letter)
             session.commit()
+            
+            # Pinecone 서비스를 이용하여 벡터를 저장
             result = embed_letter_pinecone(letter)
             print(f"save_letter result : {result}")
             return "Letter saved successfully."
@@ -53,4 +49,4 @@ def save_letter(letter: Letter):
     
 
 def get_a_letter(letter_id: int):
-    return get_letter_by_id(letter_id)
+    return get_a_letter(letter_id)
